style: separate api_request code contained in api_client
