+++ lib/src/api_client.dart
@@ -1,6 +1,4 @@
-import 'dart:convert';
-import 'dart:developer';
-
+import 'package:adcio_placement/src/api_request.dart';
 import 'package:adcio_placement/src/api_result.dart';
 import 'package:adcio_placement/src/error.dart';
 import 'package:flutter/services.dart';
@@ -9,7 +7,7 @@ import 'package:http/http.dart';
 class ApiClient {
   ApiClient({String? baseUrl}) : _baseUrl = baseUrl ?? 'https://api.adcio.ai';
 
-  final _ApiRequest _request = _ApiRequest(Client());
+  final ApiRequest _request = ApiRequest(Client());
   final String _baseUrl;
 
   Future<AdcioSuggestionRawData> suggestion({
@@ -33,11 +31,9 @@ class ApiClient {
       params['customerId'] = customerId;
     }
     params['placementId'] = placementId;
-
     if (fromAgent != null) {
       params['fromAgent'] = fromAgent;
     }
-
     if (birthYear != null) {
       params['birthYear'] = birthYear;
     }
@@ -55,7 +51,7 @@ class ApiClient {
     }
 
     final response = await _request(
-      method: _RequestMethod.post,
+      method: RequestMethod.post,
       url: url,
       params: params,
     );
@@ -81,97 +77,5 @@ class ApiClient {
         },
       );
     }
-  }
-}
-
-enum _RequestMethod { get, post }
-
-class _ApiRequest {
-  _ApiRequest(Client client) : _client = client;
-
-  final Client _client;
-
-  Future<_ApiResponse> call({
-    required _RequestMethod method,
-    Map<String, String>? headers = const {'content-type': 'application/json'},
-    required String url,
-    Map<String, dynamic>? params,
-  }) async {
-    log('[req] $method $url (${params ?? ''})');
-    final httpResponse = await _requestHttp(method, headers, url, params);
-    final response = _ApiResponse.fromHttpResponse(httpResponse);
-    log('[res] $method $url ${response.statusCode} ${response.body}');
-    return response;
-  }
-
-  Future<Response> _requestHttp(
-    _RequestMethod method,
-    Map<String, String>? headers,
-    String url,
-    Map<String, dynamic>? params,
-  ) {
-    Future<Response> result;
-    switch (method) {
-      case _RequestMethod.get:
-        if (params?.isNotEmpty ?? false) {
-          url += '?${_queryParameters(params!)}';
-        }
-        result = _client.get(
-          Uri.parse(url),
-          headers: headers,
-        );
-        break;
-      case _RequestMethod.post:
-        result = _client.post(
-          Uri.parse(url),
-          headers: headers,
-          body: json.encode(params),
-        );
-
-        break;
-    }
-
-    return result;
-  }
-
-  String _queryParameters(Map<String, dynamic> params) {
-    return params.entries.map((json) {
-      if (json.value is! Iterable) {
-        return '${json.key}=${json.value}';
-      }
-
-      return (json.value as Iterable)
-          .map((value) => '${json.key}=$value')
-          .join('&');
-    }).join('&');
-  }
-}
-
-class _ApiResponse {
-  _ApiResponse({required this.statusCode});
-
-  int statusCode;
-  dynamic body;
-
-  factory _ApiResponse.fromHttpResponse(Response res) {
-    final response = _ApiResponse(statusCode: res.statusCode);
-
-    if (res.body.isNotEmpty) {
-      final decodeData = utf8.decode(res.bodyBytes);
-      response.body = jsonDecode(decodeData);
-    }
-    return response;
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      'statusCode': statusCode,
-      'body': body,
-    };
-  }
-
-  @override
-  String toString() {
-    return toJson().toString();
   }
 }
